fix: populate timeline from entities store, not from query
